Alias `cl-format` as `format`; add `fmtstr`.
Preferring the flexibility of Common Lisp's `format` and finding it more
deserving of the name than Java's more prosaic version, it is reinstated as
such in Familiar.  A slightly more convenient version for the not-uncommon case
of using it to format strings that are not writer-bound has also been added.