Explicitly group let bindings and cond* clauses.
Indentation and whitespace is Lisps is very important not for the human reader,
if not the Lisp reader.  Indentation provides the visual cues needed to infer
the correct grouping of related information: arguments, bodies, bindings,
clauses, etc.

An early decision was made in Clojure to eliminate "unecessary" parentheses of
the Common Lisp-style let and cond forms to make the language more palatable to
those detractors of Lisps who are intolerant of the parentheses.

I posit that decision has negatively impacted the readability of the language in
practice.  I'm increasingly seeing a tendency to introduce vertical whitespace
into these forms, in an attempt to reintroduce the visual cues for the grouping
of bindings and clauses that would otherwise have been provided by conventional
indentation.  This style of using seq forms for grouping is used in Clojure in
the `letfn` form, but whether this is an intentional acknowledgement that this
type of explicit grouping is beneficial for larger forms, or whether it is an
oversight from the original decision to strip parens is unknown.

This experimental (because it is likely to be very controversial, and so is
being explicitly flagged) change reintroduces the Common Lisp conventions for
theese form to facilitate indentation-driven inference of grouping.