Introduce alpha support for testing sequences of function calls.
It is not uncommon when using external APIs to make a series of calls through
the same function (consider: `aws/invoke` in Cognitect's aws-api).  In these
cases, `with-redefs` isn't on its own sufficient, because we want different
responses to be returned for each invocation.

Although projects like [shrubbery] provide a technically superior solution to
this problem, there are times when it is necesary to use something like what is
presented here as an interim measure until the code can be refactored into
something that can better support [shrubbery]'s approach.

[shrubbery]: https://github.com/bguthrie/shrubbery